+++ application/controllers/system.php
@@ -12,7 +12,22 @@ class System extends Controller {
 	
 	function index()
 	{
+		$tabs = array(
+			array(
+				'selected' => true,
+				'img' => 'img/system.png',
+				'link' => site_url('System'),
+				'name' => 'System'
+			)
+		);
 		
+		$vars = array(
+			'tabs' => $tabs,
+			'title' => 'System Pages',
+			'page_title' => 'System Pages'
+		);
+		$this->load->vars($vars);
+		$this->load->view('sw/system/index');
 	}
 	
 	function page_list()
@@ -30,7 +45,7 @@ class System extends Controller {
 			array(
 				'selected' => true,
 				'img' => 'img/info.png',
-				'link' => site_url('System/Page List'),
+				'link' => site_url('System/Page_List'),
 				'name' => 'Page List'
 			)
 		);
@@ -43,5 +58,46 @@ class System extends Controller {
 		);
 		$this->load->vars($vars);
 		$this->load->view('sw/system/page_list');
+	}
+	
+	function user_list()
+	{
+		$users = $this->ion_auth_model->get_users();
+		
+		$tabs = array(
+			array(
+				'selected' => false,
+				'img' => 'img/system.png',
+				'link' => site_url('System'),
+				'name' => 'System'
+			),
+			array(
+				'selected' => true,
+				'img' => 'img/info.png',
+				'link' => site_url('System/User_List'),
+				'name' => 'User List'
+			)
+		);
+		
+		$vars = array(
+			'tabs' => $tabs,
+			'title' => 'User List',
+			'page_title' => 'User list',
+			'users' => $users
+		);
+		$this->load->vars($vars);
+		$this->load->view('sw/system/user_list');
+	}
+	
+	function ajax_toolbox_update()
+	{
+		if(!$this->ion_auth->logged_in() || $this->input->post('show') === FALSE)
+		{
+			echo 'failed';
+			return;
+		}
+		
+		$this->session->set_userdata('show_toolbox', $this->input->post('show'));
+		$this->ion_auth_model->update_user(array('show_toolbox' => $this->input->post('show')));
 	}
 }