+++ application/views/sw/page/edit_headinclude.php
@@ -1,2 +1,1 @@
-		<script type="text/javascript" src="<?php echo base_url(); ?>js/jquery-1.4.2.min.js"></script>
 		<script type="text/javascript" src="<?php echo base_url(); ?>js/editor.js"></script>