+++ application/views/sw/top_logged_in.php
@@ -1,8 +1,11 @@
 			<div class="container" id="top">
-				Hello, <?php echo $username; ?> (<a href="<?php echo site_url('auth/logout'); ?>">Logout</a>)
+				Hello, <a href="<?php echo site_url('User/'.$username); ?>"><?php echo $username; ?></a> (<a href="<?php echo site_url('auth/logout'); ?>">Logout</a>)
 				<?php if(isset($home_link) && $home_link === TRUE): ?>
 				<span style="float: left;">
-					<a href="<?php echo base_url(); ?>">&laquo; Go home</a>
-				</span>
+					<a href="<?php echo base_url(); ?>">&laquo; Go home</a><?php 
+					if(isset($top_extra)): ?>
+					<?php echo $top_extra; ?>
+					<?php endif; 
+				?></span>
 				<?php endif; ?>
 			</div>