+++ syncwiki.sql
@@ -147,6 +147,16 @@ CREATE TABLE `users` (
   PRIMARY KEY (`id`)
 );
 
+DROP TABLE IF EXISTS `sessions`;
+
+CREATE TABLE IF NOT EXISTS  `sessions` (
+session_id varchar(40) DEFAULT '0' NOT NULL,
+ip_address varchar(16) DEFAULT '0' NOT NULL,
+user_agent varchar(50) NOT NULL,
+last_activity int(10) unsigned DEFAULT 0 NOT NULL,
+user_data text NOT NULL,
+PRIMARY KEY (session_id)
+);